upd:final

+++ internal/domain/response.go
@@ -3,6 +3,6 @@ package domain
 type ApiResponse struct {
 	City string `json:"name"`
 	Main struct {
-		Kelvin float64 `json:"temp"`
+		Kelvin float32 `json:"temp"`
 	} `json:"main"`
 }
